fix: 🐛 fixed bugs arisen from combining main into feature branch
